Rewrite the CPU schedule to follow Mullapudi2016's advice
This schedule loosely follows the advice of the autoscheduler.

The Taylor window stuff is mostly inaccessible to the schedule,
but some basic parallelism does help.  Further visibility would
allow much better improvements to the Taylor stuff; the
autoscheduler reduces that whole section to around 5ms.

Do the norms using updates, rather than the norm() util function,
so that we can apply the same scheduling directives to them that
the autoscheduler does.