cmake: can't use Halide::Halide with try_compile for cmake >= 3.18
The problem actually appears to be that Halide's HalideConfig.cmake behaves
differently for versions <3.18 than >=3.18.
We'll let the macro caller specify which library to link against.
Using Halide::shared::Halide seems to be sufficient for both cmake versions
for now, but assumes that the shared lib is available, of course.