ImgPlane: don't set_distributed on buffers
It might be possible to use smaller buffers in a distributed schedule, but
this would also require knowledge of the schedule's partitioning. It could
save memory usage; it might save compute time if our data allocations are
more intelligent than Halide's inference, but we don't currently have
evidence to support that.
So for now we'll just remove set_distributed since it's not being used
intelligently anyway.